Desktop/devops/rag-pipeline/av_genai_chatbot/requirements.txt
gradio==5.49.1
boto3==1.40.59
python-dotenv==1.1.1

llama-index
llama-index-llms-openrouter
llama-index-retrievers-bedrock
